refactor: service 중심 구조로 service와 repository 재정리
