+++ react_web_app/src/repositories/boxRepository.ts
@@ -3,22 +3,23 @@ import { db } from '../config/firebase';
 import { BoxInfo } from '../types/box';
 
 export class BoxRepository {
+  /**
+   * 박스 문서를 조회합니다.
+   *
+   * @param boxName 박스 이름
+   * @returns 박스 정보 또는 `null`
+   */
   static async getBoxInfo(boxName: string): Promise<BoxInfo | null> {
-    try {
-      const snap = await getDoc(doc(db, 'box', boxName));
-      return snap.exists() ? (snap.data() as BoxInfo) : null;
-    } catch (error) {
-      console.error('Error getting box info:', error);
-      throw new Error('박스 정보를 불러오는데 실패했습니다.');
-    }
+    const snap = await getDoc(doc(db, 'box', boxName));
+    return snap.exists() ? (snap.data() as BoxInfo) : null;
   }
 
-  static async updateBoxInfo(boxInfo: BoxInfo): Promise<void> {
-    try {
-      await setDoc(doc(db, 'box', boxInfo.boxName), boxInfo);
-    } catch (error) {
-      console.error('Error updating box info:', error);
-      throw new Error('박스 정보 수정에 실패했습니다.');
-    }
+  /**
+   * 박스 문서를 저장합니다.
+   *
+   * @param boxInfo 저장할 박스 정보
+   */
+  static async saveBoxInfo(boxInfo: BoxInfo): Promise<void> {
+    await setDoc(doc(db, 'box', boxInfo.boxName), boxInfo);
   }
 }
