+++ react_web_app/src/repositories/memberRepository.ts
@@ -1,25 +1,18 @@
 import {
-  getDocs,
-  collection,
-  doc,
-  deleteDoc,
-  updateDoc,
   addDoc,
-  query,
-  where,
-  getDoc,
-  setDoc,
+  collection,
+  deleteDoc,
   deleteField,
-  Timestamp
+  doc,
+  getDoc,
+  getDocs,
+  query,
+  setDoc,
+  updateDoc,
+  where
 } from 'firebase/firestore';
 import { db } from '../config/firebase';
-import { Member, MemberLockerHistory } from '../types/member';
 import { UserMembership } from '../types/membership';
-import {
-  convertMembershipsFromFirebase,
-  categorizeMemberships,
-  buildMembershipInfo
-} from '../models/memberModel';
 
 export interface FirebaseMemberData {
   email: string;
@@ -32,322 +25,137 @@ export interface FirebaseMemberData {
   futureMemberships: UserMembership[];
 }
 
+export interface FirebaseMemberDocument {
+  id: string;
+  data: any;
+}
+
 export class MemberRepository {
-  static async getMembers(box: string): Promise<Member[]> {
-    try {
-      const snap = await getDocs(collection(db, `/box/${box}/member`));
-      const members: Member[] = [];
-
-      snap.forEach(docSnap => {
-        const data = docSnap.data() as any;
-
-        // 레거시 birth 필드 지원
-        if (data.birth && !data.birthDate) data.birthDate = data.birth;
-
-        const allMemberships = convertMembershipsFromFirebase(data.memberships || []);
-        const { pastMemberships, currentMemberships, futureMemberships, refundedMemberships } =
-          categorizeMemberships(allMemberships);
-
-        members.push({
-          ...data,
-          futureMemberships,
-          membershipInfo: buildMembershipInfo(pastMemberships, currentMemberships, futureMemberships, refundedMemberships)
-        });
-      });
-
-      return members;
-    } catch (error) {
-      console.error('Error fetching members:', error);
-      throw error;
-    }
+  /**
+   * 박스의 전체 회원 문서를 조회합니다.
+   *
+   * @param box 박스 이름
+   * @returns 회원 문서 목록
+   */
+  static async getMemberDocuments(box: string): Promise<FirebaseMemberDocument[]> {
+    const snap = await getDocs(collection(db, `/box/${box}/member`));
+    return snap.docs.map((docSnap) => ({ id: docSnap.id, data: docSnap.data() }));
   }
 
+  /**
+   * 특정 회원 문서를 조회합니다.
+   *
+   * @param box 박스 이름
+   * @param email 회원 이메일
+   * @returns 회원 문서 또는 `null`
+   */
+  static async getMemberDocument(box: string, email: string): Promise<any | null> {
+    const snap = await getDoc(doc(db, `/box/${box}/member`, email));
+    return snap.exists() ? snap.data() : null;
+  }
+
+  /**
+   * 회원 문서를 삭제합니다.
+   *
+   * @param box 박스 이름
+   * @param email 회원 이메일
+   */
   static async deleteMember(box: string, email: string): Promise<void> {
-    try {
-      await deleteDoc(doc(db, `/box/${box}/member`, email));
-    } catch (error) {
-      console.error('Error deleting member:', error);
-      throw error;
-    }
+    await deleteDoc(doc(db, `/box/${box}/member`, email));
   }
 
-  static async updateMemberMembership(box: string, email: string, membershipData: any): Promise<void> {
-    try {
-      await updateDoc(doc(db, `/box/${box}/member`, email), membershipData);
-    } catch (error) {
-      console.error('Error updating member membership:', error);
-      throw error;
-    }
+  /**
+   * 회원 문서를 부분 수정합니다.
+   *
+   * @param box 박스 이름
+   * @param email 회원 이메일
+   * @param payload 수정할 필드
+   */
+  static async updateMember(box: string, email: string, payload: any): Promise<void> {
+    await updateDoc(doc(db, `/box/${box}/member`, email), payload);
   }
 
+  /**
+   * 회원 문서를 병합 저장합니다.
+   *
+   * @param box 박스 이름
+   * @param email 회원 이메일
+   * @param payload 저장할 데이터
+   */
+  static async setMember(box: string, email: string, payload: any): Promise<void> {
+    await setDoc(doc(db, `/box/${box}/member`, email), payload, { merge: true });
+  }
+
+  /**
+   * 자동 생성 ID로 회원 문서를 추가합니다.
+   *
+   * @param box 박스 이름
+   * @param memberData 저장할 회원 데이터
+   */
   static async addMember(box: string, memberData: FirebaseMemberData): Promise<void> {
-    try {
-      await addDoc(collection(db, `/box/${box}/member`), memberData);
-    } catch (error) {
-      console.error('Error adding member:', error);
-      throw error;
-    }
+    await addDoc(collection(db, `/box/${box}/member`), memberData);
   }
 
-  static async searchMembersByName(box: string, searchName: string): Promise<Member[]> {
-    try {
-      const all = await this.getMembers(box);
-      const term = searchName.trim().toLowerCase();
-      if (!term) return all;
-      return all.filter(m => m.realName.toLowerCase().includes(term));
-    } catch (error) {
-      console.error('Error searching members:', error);
-      throw error;
-    }
+  /**
+   * 특정 필드 값과 일치하는 사용자 문서를 조회합니다.
+   *
+   * @param field 조회 필드명
+   * @param value 조회 값
+   * @returns 사용자 데이터 목록
+   */
+  static async getUsersByField(field: string, value: string): Promise<any[]> {
+    const snap = await getDocs(query(collection(db, '/user'), where(field, '==', value)));
+    return snap.docs.map((docSnap) => docSnap.data());
   }
 
-  static async assignLockerToMember(
-    box: string,
-    email: string,
-    lockerNumber: number,
-    startDate: string,
-    endDate: string,
-    key: string,
-    price: string,
-    paymentType: 'cash' | 'card'
-  ): Promise<void> {
-    try {
-      const memberRef = doc(db, `/box/${box}/member`, email);
-      const memberDoc = await getDoc(memberRef);
-
-      let lockerHistory: MemberLockerHistory[] = [];
-      if (memberDoc.exists()) {
-        lockerHistory = (memberDoc.data().lockerHistory || []) as MemberLockerHistory[];
-      }
-
-      lockerHistory.push({
-        lockerNum: lockerNumber,
-        startDate,
-        endDate,
-        createdAt: Timestamp.now(),
-        key,
-        price,
-        paymentType
-      });
-
-      await updateDoc(memberRef, { lockerHistory });
-    } catch (error) {
-      console.error('Error assigning locker to member:', error);
-      throw error;
+  /**
+   * 이메일과 일치하는 사용자 문서를 부분 수정합니다.
+   *
+   * @param email 사용자 이메일
+   * @param userData 수정할 데이터
+   * @returns 수정 데이터 또는 `null`
+   */
+  static async updateUsersByEmail(email: string, userData: any): Promise<any> {
+    const snap = await getDocs(query(collection(db, '/user'), where('email', '==', email)));
+    if (snap.empty) {
+      return null;
     }
+
+    for (const docSnap of snap.docs) {
+      await updateDoc(docSnap.ref, userData);
+    }
+
+    return userData;
   }
 
-  static async updateLockerHistoryEndDate(
-    box: string,
-    email: string,
-    key: string,
-    newEndDate: string
-  ): Promise<void> {
-    try {
-      const memberRef = doc(db, `/box/${box}/member`, email);
-      const memberDoc = await getDoc(memberRef);
-
-      if (!memberDoc.exists()) throw new Error('회원 정보를 찾을 수 없습니다.');
-
-      const lockerHistory: any[] = memberDoc.data().lockerHistory || [];
-      const idx = lockerHistory.findIndex(item => item.key === key);
-
-      if (idx === -1) {
-        console.warn(`락커 히스토리를 찾을 수 없습니다. email: ${email}, key: ${key}`);
-        return;
-      }
-
-      lockerHistory[idx] = { ...lockerHistory[idx], endDate: newEndDate };
-      await updateDoc(memberRef, { lockerHistory });
-    } catch (error) {
-      console.error('Error updating locker history endDate:', error);
-      throw error;
-    }
+  /**
+   * 가입 신청 문서를 조회합니다.
+   *
+   * @param boxName 박스 이름
+   * @returns 신청 데이터 맵 또는 `null`
+   */
+  static async getApplicantMap(boxName: string): Promise<Record<string, any> | null> {
+    const snap = await getDoc(doc(db, `box/${boxName}/applied/applieddoc`));
+    return snap.exists() ? (snap.data() as Record<string, any>) : null;
   }
 
-  static async unassignLockerFromMember(
-    box: string,
-    email: string,
-    lockerNumber: number,
-    endDate: string,
-    key: string
-  ): Promise<void> {
-    try {
-      const memberRef = doc(db, `/box/${box}/member`, email);
-      const memberDoc = await getDoc(memberRef);
-
-      if (!memberDoc.exists()) throw new Error('회원 정보를 찾을 수 없습니다.');
-
-      const lockerHistory: any[] = memberDoc.data().lockerHistory || [];
-      const idx = lockerHistory.findIndex(item => item.key === key && item.lockerNum === lockerNumber);
-
-      if (idx === -1) throw new Error('해당 락커 할당 기록을 찾을 수 없습니다.');
-
-      lockerHistory[idx] = { ...lockerHistory[idx], endDate };
-      await updateDoc(memberRef, { lockerHistory });
-    } catch (error) {
-      console.error('Error unassigning locker from member:', error);
-      throw error;
-    }
+  /**
+   * 가입 신청 문서에서 특정 신청 정보를 제거합니다.
+   *
+   * @param email 신청자 이메일
+   * @param boxName 박스 이름
+   */
+  static async deleteApplication(email: string, boxName: string): Promise<void> {
+    await setDoc(doc(db, `box/${boxName}/applied/applieddoc`), { [email]: deleteField() }, { merge: true });
   }
 
-  static async getUserByEmail(email: string): Promise<any> {
-    try {
-      const snap = await getDocs(query(collection(db, '/user'), where('email', '==', email)));
-      let userData = null;
-      snap.forEach(d => { userData = d.data(); });
-      return userData;
-    } catch (error) {
-      console.error('Error fetching user by email:', error);
-      throw error;
-    }
-  }
-
-  static async getUserByPhone(phone: string): Promise<any> {
-    try {
-      const snap = await getDocs(query(collection(db, '/user'), where('phone', '==', phone)));
-      let userData = null;
-      snap.forEach(d => { userData = d.data(); });
-      return userData;
-    } catch (error) {
-      console.error('Error fetching user by phone:', error);
-      throw error;
-    }
-  }
-
-  static async getUserByRealName(realName: string): Promise<any[]> {
-    try {
-      const snap = await getDocs(query(collection(db, '/user'), where('realName', '==', realName)));
-      const users: any[] = [];
-      snap.forEach(d => users.push(d.data()));
-      return users;
-    } catch (error) {
-      console.error('Error fetching users by realName:', error);
-      throw error;
-    }
-  }
-
-  static async getUserByNickName(nickName: string): Promise<any[]> {
-    try {
-      const snap = await getDocs(query(collection(db, '/user'), where('nickName', '==', nickName)));
-      const users: any[] = [];
-      snap.forEach(d => users.push(d.data()));
-      return users;
-    } catch (error) {
-      console.error('Error fetching users by nickName:', error);
-      throw error;
-    }
-  }
-
-  static async createMember(box: string, memberData: any): Promise<void> {
-    try {
-      const memberDocRef = doc(collection(db, `/box/${box}/member`), memberData.email);
-      const snap = await getDoc(memberDocRef);
-      if (snap.exists()) {
-        console.log(`Member with email ${memberData.email} already exists. Skipping creation.`);
-        return;
-      }
-      await setDoc(memberDocRef, memberData);
-    } catch (error) {
-      console.error('멤버 추가 중 오류 발생:', error);
-      throw error;
-    }
-  }
-
-  static async updateUser(email: string, userData: any): Promise<any> {
-    try {
-      const snap = await getDocs(query(collection(db, '/user'), where('email', '==', email)));
-      if (snap.empty) {
-        console.warn('해당 이메일로 사용자를 찾을 수 없습니다:', email);
-        return null;
-      }
-      for (const docSnap of snap.docs) {
-        await updateDoc(docSnap.ref, userData);
-      }
-      return userData;
-    } catch (error) {
-      console.error('사용자 업데이트 중 오류 발생:', error);
-      throw error;
-    }
-  }
-
-  static async fetchApplicants(boxName: string): Promise<any[]> {
-    try {
-      const snap = await getDoc(doc(db, `box/${boxName}/applied/applieddoc`));
-      const applicants: any[] = [];
-      if (snap.exists()) {
-        const data = snap.data();
-        for (const email in data) {
-          if (Object.prototype.hasOwnProperty.call(data, email)) {
-            const a = data[email];
-            applicants.push({
-              name: a.realName || '',
-              email: a.email || '',
-              phone: a.phone || '',
-              boxName,
-              birth: a.birth || undefined
-            });
-          }
-        }
-      }
-      return applicants;
-    } catch (error) {
-      console.error('Failed to fetch applicants:', error);
-      return [];
-    }
-  }
-
-  static async approveApplicant(email: string, boxName: string): Promise<void> {
-    try {
-      const userDoc = await this.getUserByEmail(email);
-      if (!userDoc) throw new Error('사용자 정보를 찾을 수 없습니다.');
-      if (!userDoc.boxName?.startsWith('?')) throw new Error('유효하지 않은 신청 상태입니다.');
-
-      const actualBoxName = userDoc.boxName.slice(1);
-      await this.removeApplication(email, actualBoxName);
-
-      if (userDoc.hasOwnProperty('memberships')) delete userDoc.memberships;
-      if (userDoc.birth && !userDoc.birthDate) {
-        userDoc.birthDate = userDoc.birth;
-        delete userDoc.birth;
-      }
-
-      userDoc.boxName = actualBoxName;
-      userDoc.joinedAt = Timestamp.now();
-
-      await this.createMember(actualBoxName, userDoc);
-    } catch (error) {
-      console.error('Failed to approve applicant:', error);
-      throw error;
-    }
-  }
-
-  static async rejectApplicant(email: string, boxName: string): Promise<void> {
-    try {
-      await setDoc(doc(db, `box/${boxName}/applied/applieddoc`), { [email]: deleteField() }, { merge: true });
-      await updateDoc(doc(db, `user/${email}`), { boxName: '' });
-    } catch (error) {
-      console.error('Failed to reject applicant:', error);
-      throw error;
-    }
-  }
-
-  static async removeApplication(email: string, boxName: string): Promise<void> {
-    try {
-      await setDoc(doc(db, `box/${boxName}/applied/applieddoc`), { [email]: deleteField() }, { merge: true });
-      await updateDoc(doc(db, `user/${email}`), { boxName });
-    } catch (error) {
-      console.error('Failed to remove application:', error);
-      throw error;
-    }
-  }
-
-  static async updateMemberMemo(box: string, email: string, memo: string): Promise<void> {
-    try {
-      await updateDoc(doc(db, `/box/${box}/member`, email), { memo });
-    } catch (error) {
-      console.error('Error updating member memo:', error);
-      throw error;
-    }
+  /**
+   * 사용자 문서의 박스 이름을 수정합니다.
+   *
+   * @param email 사용자 이메일
+   * @param boxName 반영할 박스 이름
+   */
+  static async updateUserBoxName(email: string, boxName: string): Promise<void> {
+    await updateDoc(doc(db, `user/${email}`), { boxName });
   }
 }
